+++ satella/__init__.py
@@ -1,3 +1,2 @@
 # coding=UTF-8
-__version__ = '2.1.12a1'
-
+__version__ = '2.1.12rc1'