+++ satella/coding/concurrent/__init__.py
@@ -1,5 +1,6 @@
 from .callablegroup import *
 from .monitor import *
+from .thread import TerminableThread
 from .locked_dataset import LockedDataset
 
-__all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup']
+__all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread']